requirements.txt
arrow==1.2.3
asgiref==3.7.2
binaryornot==0.4.4
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cookiecutter==2.1.1
databases==0.7.0
Django==4.2.2
django-browser-reload==1.10.0
django-tailwind==3.5.0
greenlet==2.0.2
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.3
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4.48
sqlite-orm==0.0.2b0
sqlparse==0.4.4
text-unidecode==1.3
typesystem==0.3.1
typing_extensions==4.6.3
urllib3==2.0.3
